requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.5
<<<<<<< HEAD
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.3.1
=======
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.3.1
psycopg2==2.9.1
>>>>>>> b54bbea... modified files
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.3.0
